// frontend/src/app/generated/postProject.ts
/**
 * RESTful API of iHRM
 * The RESTful API of the main core of the iHRM application. Add-ons will have their own API documentation.
 *
 * OpenAPI spec version: 2020.06.14
 * Contact: ihrm@imbus.tn
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface PostProject { 
    customerId: number;
    name: string;
    description?: string;
    note?: string;
    isActive?: boolean;
}